Implement Basic Battlenet

+++ src/world/Server/WorldSession.h
@@ -227,7 +227,7 @@ public:
 
 #if VERSION_STRING != Cata
         uint32_t GetClientBuild() { return client_build; }
-        void SetClientBuild(uint32_t build) { client_build = static_cast<uint16_t>(build); }
+        void SetClientBuild(uint32_t build) { client_build = build; }
 #else
         uint16_t GetClientBuild() { return client_build; }
         void SetClientBuild(uint16_t build) { client_build = build; }
